Use same allocation method for allocating all bit vectors to ensure padding and guard bits.
