+++ searchlib/src/vespa/searchlib/common/CMakeLists.txt
@@ -6,6 +6,7 @@ vespa_add_library(searchlib_common OBJECT
     bitvector.cpp
     bitvectorcache.cpp
     bitvectoriterator.cpp
+    bitword.cpp
     condensedbitvectors.cpp
     documentlocations.cpp
     documentsummary.cpp
