+++ searchlib/src/vespa/searchlib/common/partialbitvector.cpp
@@ -1,8 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-// Copyright (C) 1998-2003 Fast Search & Transfer ASA
-// Copyright (C) 2003 Overture Services Norway AS
 
-#include <vespa/fastos/fastos.h>
 #include "partialbitvector.h"
 
 namespace search {
@@ -11,7 +8,7 @@ using vespalib::alloc::Alloc;
 
 PartialBitVector::PartialBitVector(Index start, Index end) :
     BitVector(),
-    _alloc(Alloc::alloc(numActiveBytes(start, end), 0x1000000, 0x1000))
+    _alloc(allocatePaddedAndAligned(start, end))
 {
     init(_alloc.get(), start, end);
     clear();
@@ -19,14 +16,13 @@ PartialBitVector::PartialBitVector(Index start, Index end) :
 
 PartialBitVector::PartialBitVector(const BitVector & org, Index start, Index end) :
         BitVector(),
-        _alloc(Alloc::alloc(numActiveBytes(start, end), 0x1000000, 0x1000))
+        _alloc(allocatePaddedAndAligned(start, end))
 {
     init(_alloc.get(), start, end);
-    memcpy(_alloc.get(), org.getWordIndex(start), _alloc.size());
+    memcpy(_alloc.get(), org.getWordIndex(start), numActiveBytes(start, end));
+    setBit(size());
 }
 
-PartialBitVector::~PartialBitVector()
-{
-}
+PartialBitVector::~PartialBitVector() { }
 
 } // namespace search
